+++ src/core/engine/index.ts
@@ -16,3 +16,4 @@ export * from './design-to-estimate';
 export * from './fuel';
 export * from './motion';
 export * from './build-cost';
+export * from './throughput';